requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
Django==5.2.4
djangorestframework==3.16.0
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
selenium==4.34.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
